Rename `appuser` to `dbowner`
As noted in #215, the naming `appuser` may confuse users into
thinking the account to use for DATABASE_URL is the same as the
account that should be used by the application to connect to the
database. While this may be true in some setups, it is not a hard
requirement (see the discussion in #215 for further details).

resolves #215